+++ samples/botframework-webchat-control/src/interfaces/IResultAction.ts
@@ -1,6 +1,0 @@
-import { IWebChatAction } from "./IWebChatAction";
-
-export interface IResultAction {
-    dispatchAction: IWebChatAction | null;
-    nextAction: IWebChatAction;
-}