+++ samples/botframework-webchat-control/src/interfaces/IWebChatAction.ts
@@ -1,4 +1,0 @@
-export interface IWebChatAction {
-    type: string;
-    payload: object;
-}