+++ samples/react-chat-widget-control/src/components/ChatWidget/ChatWidget.test.tsx
@@ -1,3 +1,0 @@
-test('Test', () => {
- expect(true).toBe(true);
-});